init: add config files
