+++ index.ts
@@ -1,0 +1,7 @@
+import { httpServer } from './src/http_server';
+import './src/websocket_server';
+
+const HTTP_PORT = 8181;
+
+console.log(`Start static http server on the ${HTTP_PORT} port!`);
+httpServer.listen(HTTP_PORT);
